updated token variable type to long in all instances

+++ src/interfaces/ExamServer.java
@@ -16,15 +16,15 @@ public interface ExamServer extends Remote {
 		UnauthorizedAccess, RemoteException;
 
 	// Return a summary list of Assessments currently available for this studentid
-	public List<String> getAvailableSummary(int token, int studentid) throws
+	public List<String> getAvailableSummary(long token, int studentid) throws
 		UnauthorizedAccess, NoMatchingAssessment, RemoteException;
 
 	// Return an Assessment object associated with a particular course code
-	public Assessment getAssessment(int token, int studentid, String courseCode) throws
+	public Assessment getAssessment(long token, int studentid, String courseCode) throws
 		UnauthorizedAccess, NoMatchingAssessment, RemoteException;
 
 	// Submit a completed assessment
-	public void submitAssessment(int token, int studentid, Assessment completed) throws 
+	public void submitAssessment(long token, int studentid, Assessment completed) throws 
 		UnauthorizedAccess, NoMatchingAssessment, RemoteException;
 
 }
